add error exclusion to vpnd
